% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmaps.R
\name{dmaps}
\alias{dmaps}
\title{<Add Title>}
\usage{
dmaps(data = NULL, mapName, opts = NULL, regionCols = NULL,
  codeCol = NULL, groupCol = NULL, valueCol = NULL, regions = NULL,
  bubbles = NULL, width = "100\%", height = "100\%", ...)
}
\description{
<Add Description>
}
