% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dmaps.R
\name{dmaps}
\alias{dmaps}
\title{<Add Title>}
\usage{
dmaps(mapName, data = NULL, groupCol = NULL, valueCol = NULL,
  regionCols = NULL, bubbles = NULL, opts = NULL, width = "100\%",
  height = "80\%", ...)
}
\description{
<Add Description>
}

