% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmaps.R
\name{dmaps}
\alias{dmaps}
\title{<Add Title>}
\usage{
dmaps(mapName, data = NULL, regionCols = NULL, codeCol = NULL,
  groupCol = NULL, valueCol = NULL, bubbles = NULL, opts = NULL,
  width = "100\%", height = "80\%", ...)
}
\description{
<Add Description>
}

